+++ Source/Engine/Animators/Nodes/AnimationClipNodeBase.h
@@ -18,10 +18,10 @@ public:
 	void setLooping(bool value);
 	bool getStitchFinalFrame() { return m_stitchFinalFrame; }
 	void setStitchFinalFrame(bool value);
-	unsigned int getTotalDuration();
+	float getTotalDuration();
 	Vector3D& getTotalDelta();
 	unsigned int getLastFrame();
-	std::vector<unsigned int>& getDurations() { return m_durations; }
+	std::vector<float>& getDurations() { return m_durations; }
 
 protected:
 	virtual void updateStitch();
@@ -33,10 +33,10 @@ protected:
 	bool m_looping;
 	bool m_stitchDirty;
 	bool m_stitchFinalFrame;
-	unsigned int m_totalDuration;
+	float m_totalDuration;
 	unsigned int m_lastFrame;
 	unsigned int m_numFrames;
-	std::vector<unsigned int> m_durations;
+	std::vector<float> m_durations;
 	Vector3D m_totalDelta;
 };
